Implemented Entity clone in Go (enh. #967)
