--- sdk/go/core/ydk/types/types_test.go
@@ -1,11 +1,0 @@
-package types
-
-import "testing"
-
-func TestGetRelativeEntityPath(t *testing.T) {
-
-}
-
-func TestYLeaf_GetNameLeafdata(t *testing.T) {
-
-}
